backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "physical-ai-vla-backend"
version = "0.1.0"
description = "Backend API for the Vision-Language-Action (VLA) & Capstone module of the Physical AI & Humanoid Robotics textbook"
readme = "README.md"
authors = [{name = "AI Development Team", email = "dev@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-dotenv==1.0.0",
    "openai==1.3.8",
    "google-generativeai==0.4.1",
    "qdrant-client==1.7.0",
    "PyJWT==2.8.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.23",
    "asyncpg==0.29.0",
    "pydantic==2.5.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
    "slowapi==0.1.8",
    "python-json-logger==2.0.7"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "flake8",
    "mypy",
    "isort"
]

[project.urls]
Homepage = "https://github.com/your-org/physical-ai-textbook"
Repository = "https://github.com/your-org/physical-ai-textbook.git"

[tool.setuptools.packages.find]
where = ["src"]
include = ["physical_ai_vla*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
    "*/migrations/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]